// src/app/sgi/models/location-service-constants.ts
export enum LocationServiceConstants {
    PageSize = 30,
}
export enum LocationServiceFieldOptions{
    NAME = 'name',
    ASCII_NAME = 'ascii_name',
    GEONAME_ID = 'geoname_id',
    ALTERNATE_NAMES = 'alternate_names',
    COUNTRY_CODE = 'country_code',
    COUNTRY_NAME = 'cou_name_en',
    POPULATION = 'population',
    TIMEZONE = 'timezone',
    LABEL = 'label_en'
}